Refactor images related code
 * move different formats support to its own directory
 * add common CaptureBufferToFile subroutine
